Make `cobalt-cli` usable as a library as well as an executable
